Initial Commit.
Created project where you can press space to import photos and it will render those photos on the screen. Right now it re-renders every time you resize the window or import a new photo, but that makes resizing slow, as it's not actually rendering the image as a whole but rendering each pixel as a rectangle for now. This is obviously inefficient, but we are doing this so that we can later do matrix transformations on the pixels.